requirements.txt
myenv\Scripts\activatePySide6==6.6.3
PySide6_Addons==6.6.3
PySide6_Essentials==6.6.3
shiboken6==6.6.3
